enhancement (list-view): create abstract class KuiView and extend it in list-view component
