+++ projects/knora/viewer/src/lib/view/list-view/list-view.component.ts
@@ -1,4 +1,6 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
+import { ActivatedRoute, Params, Router } from '@angular/router';
+import { Subscription } from 'rxjs';
 import {
     ApiServiceError,
     ExtendedSearchParams,
@@ -11,15 +13,14 @@ import {
     SearchParamsService,
     SearchService
 } from '@knora/core';
-import { ActivatedRoute, Params, Router } from '@angular/router';
-import { Subscription } from 'rxjs';
+import { KuiView } from '../kui-view';
 
 @Component({
     selector: 'kui-list-view',
     templateUrl: './list-view.component.html',
     styleUrls: ['./list-view.component.scss']
 })
-export class ListViewComponent implements OnDestroy, OnInit {
+export class ListViewComponent extends KuiView {
 
     KnoraConstants = KnoraConstants;
 
@@ -33,139 +34,17 @@ export class ListViewComponent implements OnDestroy, OnInit {
     searchQuery: string;
     searchMode: string;
 
+    isLoading = true;
     errorMessage: any = undefined;
     navigationSubscription: Subscription;
 
     constructor(
-        private _route: ActivatedRoute,
-        private _searchService: SearchService,
-        private _cacheService: OntologyCacheService,
-        private _searchParamsService: SearchParamsService,
-        private _gravSearchService: GravsearchGenerationService,
-        private _router: Router
+        protected _route: ActivatedRoute,
+        protected _searchService: SearchService,
+        protected _searchParamsService: SearchParamsService,
+        protected _router: Router
     ) {
-    }
-
-    ngOnInit() {
-        this.navigationSubscription = this._route.paramMap.subscribe((params: Params) => {
-            this.searchMode = params.get('mode');
-
-            // init offset to 0
-            this.offset = 0;
-            this.result = [];
-
-            if (this.searchMode === 'fulltext') {
-                this.searchQuery = params.get('q');
-            } else if (this.searchMode === 'extended') {
-                this.gravsearchGenerator = this._searchParamsService.getSearchParams();
-                this.generateGravsearchQuery();
-            }
-
-            this.rerender = true;
-            this.getResult();
-        });
-    }
-
-    ngOnDestroy() {
-        if (this.navigationSubscription !== undefined) {
-            this.navigationSubscription.unsubscribe();
-        }
-    }
-
-    /**
-     * Generates the Gravsearch query for the current offset.
-     */
-    private generateGravsearchQuery() {
-
-        const gravsearch: string | boolean = this.gravsearchGenerator.generateGravsearch(this.offset);
-        if (gravsearch === false) {
-            // no valid search params (application has been reloaded)
-            // go to root
-            this._router.navigate([''], { relativeTo: this._route });
-            return;
-        } else {
-            this.searchQuery = <string>gravsearch;
-        }
-    }
-
-    /**
-       * Get search result from Knora - 2 cases: simple search and extended search
-       */
-    getResult() {
-        // this.isLoading = true;
-
-        // FULLTEXT SEARCH
-        if (this.searchMode === 'fulltext') {
-            // perform count query
-            /* if (this.offset === 0) {
-
-                this._searchService.doFullTextSearchCountQueryCountQueryResult(this.searchQuery)
-                    .subscribe(
-                        this.showNumberOfAllResults,
-                        (error: any) => {
-                            this.errorMessage = <any>error;
-                            // console.log('numberOfAllResults', this.numberOfAllResults);
-                        }
-                    );
-            } */
-
-            // perform full text search
-            this._searchService.doFullTextSearchReadResourceSequence(this.searchQuery, this.offset)
-                .subscribe(
-                    this.processSearchResults, // function pointer
-                    (error: any) => {
-                        this.errorMessage = <any>error;
-                    }
-                );
-
-            // EXTENDED SEARCH
-        } else if (this.searchMode === 'extended') {
-            // perform count query
-            /* if (this.offset === 0) {
-                this._searchService.doExtendedSearchCountQueryCountQueryResult(this.searchQuery)
-                    .subscribe(
-                        this.showNumberOfAllResults,
-                        (error: any) => {
-                            this.errorMessage = <any>error;
-                        }
-                    );
-            } */
-            this._searchService.doExtendedSearchReadResourceSequence(this.searchQuery)
-                .subscribe(
-                    this.processSearchResults, // function pointer
-                    (error: any) => {
-                        this.errorMessage = <any>error;
-                    });
-
-        } else {
-            this.errorMessage = `search mode invalid: ${this.searchMode}`;
-        }
-    }
-
-    /**
-     *
-     * Converts search results from JSON-LD to a [[ReadResourcesSequence]] and requests information about ontology entities.
-     * This function is passed to `subscribe` as a pointer (instead of redundantly defining the same lambda function).
-     *
-     * @param {ReadResourcesSequence} searchResult the answer to a search request.
-     */
-    private processSearchResults = (searchResult: ReadResourcesSequence) => {
-
-        // assign ontology information to a variable so it can be used in the component's template
-        if (this.ontologyInfo === undefined) {
-            // init ontology information
-            this.ontologyInfo = searchResult.ontologyInformation;
-        } else {
-            // update ontology information
-            this.ontologyInfo.updateOntologyInformation(searchResult.ontologyInformation);
-        }
-        // append results to search results
-        // console.log('results 1', this.result);
-        this.result = this.result.concat(searchResult.resources);
-
-        // this.isLoading = false;
-        this.rerender = false;
-
+        super(_route, _searchService, _searchParamsService, _router);
     }
 
 }
